Add files via upload
This Project1.ipynb is the Jupyter notebook file that is my final submission. In the future I would like to rearrange the data, and try to make further conclusions regarding the education of the different counties I selected to try and understand why a county like Grant has such a high relative poverty rate, but due to the time restraint I did not.